+++ pulsar/__init__.py
@@ -36,6 +36,8 @@ from .utils.log import *
 DEFAULT_PORT = 8060
 ASYNC_TIMEOUT = None
 SERVER_SOFTWARE = "python-{0}/{1}".format(SERVER_NAME, version)
+JAPANESE = b'\xe3\x83\x91\xe3\x83\xab\xe3\x82\xb5\xe3\x83\xbc'.decode('utf-8')
+CHINESE = b'\xe8\x84\x89\xe5\x86\xb2\xe6\x98\x9f'.decode('utf-8')
 
 from .utils.exceptions import *
 from .utils import system