+++ pulsar/apps/ws/__init__.py
@@ -82,7 +82,7 @@ class GeneralWebSocket(object):
         if self.handle.match(environ):
             headers, parser = self.handle_handshake(environ, start_response)
             response = wsgi.WsgiResponse(101, (b'',), response_headers=headers)
-            upgrade = environ['upgrade_protocol']
+            upgrade = environ['connection.upgrade']
             upgrade(partial(WebSocketProtocol, self.handle, environ, parser))
             return response(environ, start_response)
     
@@ -277,7 +277,7 @@ class WebSocketProtocol(pulsar.ProtocolConsumer):
             self.transport.close()
     
     
-class WebSocketClientProtocol(pulsar.ClientProtocolConsumer):
+class WebSocketClientProtocol(pulsar.ProtocolConsumer):
     
     def __init__(self, *args):
         super(WebSocketClientProtocol, self).__init__(*args)