+++ pulsar/apps/wsgi/__init__.py
@@ -59,9 +59,4 @@ class WSGIServer(SocketServer):
     def handler(self):
         '''Build a :class:`WsgiHandler` for the WSGI callable provided
 and return an :class:`HttpResponse` factory function.'''
-        if not isinstance(self.callable, WsgiHandler):
-            hnd = self.callable
-            if not isinstance(hnd, (list, tuple)):
-                hnd = [hnd]
-            self.callable = WsgiHandler(hnd)
         return partial(HttpServerResponse, self.callable)